requirements.txt
moviepy==1.0.3
pdf2image==1.16.3
Pillow==10.2.0
customtkinter==5.2.2